refactor(patterns/form): split into 4 focused submodules
- Extract field logic to field.rs (323 lines)
- Extract state to state.rs (234 lines)
- Extract tests to tests.rs (966 lines)
- Add types.rs for shared types (19 lines)
- Original mod.rs was 1,540 lines
